list and form split out into components but not rendering properly now
